Consolidate fixtures in conftest.py
Centralizing the fixtures in conftest.py enables their reuse in other
test suites.
